Moved iterator to conduit_node_iterator.hpp
