+++ src/libs/conduit/conduit_node.hpp
@@ -80,194 +80,6 @@
 namespace conduit
 {
 
-// TODO remove when using C++11
-namespace internal {
-template<bool cond, typename T, typename F>
-struct conditional {
-    typedef T type;
-};
-
-template<typename T, typename F>
-struct conditional<false, T, F> {
-    typedef F type;
-};
-
-template<bool cond, class T = void>
-struct enable_if {};
-
-template<class T>
-struct enable_if<true, T> { typedef T type; };
-
-template<class T, class U>
-struct is_same {static const bool value = false;};
-
-template<class T>
-struct is_same<T, T> {static const bool value = true;};
-}
-
-class Node;
-
-/**
- * Class template for defining C++-style iterators for the Node class.
- *
- * The class meets the requirements for a random access iterator.
- *
- * @tparam is_const there this iterator iterates over const or non-const
- * values.
- */
-template<bool is_const>
-class CONDUIT_API node_iterator_template {
-public:
-    typedef std::random_access_iterator_tag iterator_category;
-    typedef Node value_type;
-    typedef typename internal::conditional<is_const, Node const *, Node *>::type pointer;
-    typedef typename internal::conditional<is_const, Node const &, Node &>::type reference;
-    typedef index_t difference_type;
-
-    node_iterator_template() : m_parent(NULL), m_index(0) {}
-
-    node_iterator_template(node_iterator_template const &rhs) :
-        m_parent(rhs.m_parent), m_index(rhs.m_index) {}
-
-    /**
-     * Create a new iterator pointing at the given parent node and index.
-     *
-     * @param parent the parent node whose children to iterate over
-     * @param index the index of the child to which this iterator points
-     */
-    node_iterator_template(pointer parent, index_t index=0) :
-            m_parent(parent), m_index(index) {}
-
-#if __cplusplus >= 201103L
-    // If someone has a nice, simple way to accomplish this with C++03,
-    // please change this and enable the test.
-    template <bool is_const_ = is_const, typename = typename std::enable_if<is_const_>::type>
-    node_iterator_template(node_iterator_template<false> &rhs) :
-            m_parent(rhs.get_parent()), m_index(rhs.get_index()) {}
-#endif
-
-    bool operator==(node_iterator_template const &rhs) const {
-        return m_parent == rhs.m_parent && m_index == rhs.m_index;
-    }
-
-    bool operator!=(node_iterator_template const &rhs) const {
-        return ! (*this == rhs);
-    }
-
-    node_iterator_template &operator++() {
-        ++m_index;
-        return *this;
-    }
-
-    node_iterator_template operator++(int) {
-        node_iterator_template result = *this;
-        ++*this;
-        return result;
-    }
-
-    node_iterator_template &operator--() {
-        --m_index;
-        return *this;
-    }
-
-    node_iterator_template operator--(int) {
-        node_iterator_template result = *this;
-        --*this;
-        return result;
-    }
-
-    reference operator*() {
-        return (*this)[0];
-    }
-
-    pointer operator->() {
-        return m_parent->child_ptr(m_index);
-    }
-
-    node_iterator_template& operator+=(index_t offset) {
-        m_index += offset;
-        return *this;
-    }
-
-    node_iterator_template& operator-=(index_t offset) {
-        m_index -= offset;
-        return *this;
-    }
-
-    node_iterator_template operator+(index_t rhs) const {
-        node_iterator_template result(*this);
-        result += rhs;
-        return result;
-    }
-
-    node_iterator_template operator-(index_t rhs) const {
-        node_iterator_template result(*this);
-        result -= rhs;
-        return result;
-    }
-
-    index_t operator-(node_iterator_template const &rhs) const {
-        return m_index - rhs.m_index;
-    }
-
-    reference operator[](index_t offset) const {
-        return m_parent->child(m_index + offset);
-    }
-
-    bool operator<(node_iterator_template const &rhs) const {
-        return m_index < rhs.m_index;
-    }
-
-    bool operator<=(node_iterator_template const &rhs) const {
-        return m_index <= rhs.m_index;
-    }
-
-    bool operator>(node_iterator_template const &rhs) const {
-        return m_index > rhs.m_index;
-    }
-
-    bool operator>=(node_iterator_template const &rhs) const {
-        return m_index >= rhs.m_index;
-    }
-
-    /**
-     * Get the node being iterated over (the parent).
-     *
-     * @return the parent node
-     */
-    Node const* get_parent() const {
-        return m_parent;
-    }
-
-    /**
-     * Get the index of this iterator's node in the parent node
-     *
-     * @return the index corresponding to this iterator's position in the
-     * parent
-     */
-    index_t get_index() const {
-        return m_index;
-    }
-private:
-    pointer m_parent;
-    index_t m_index;
-};
-
-/**
- * Add an offset to an iterator when the iterator is on the right-hand side.
- *
- * @tparam NodeIter the type of the iterator
- *
- * @param lhs the offset to add to the iterator
- * @param rhs the iterator to which to add the offset
- * @return the offset iterator
- */
-template<typename NodeIter>
-NodeIter operator+(index_t lhs, NodeIter const &rhs) {
-    return rhs + lhs;
-}
-
-
 //-----------------------------------------------------------------------------
 // -- forward declarations required for conduit::Node --
 //-----------------------------------------------------------------------------
